Restructure example packages
