--- src/testFixtures/java/example/configuration/FailingAfterClassConfigurationMethodTestCase.java
@@ -1,26 +1,0 @@
-/*
- * Copyright 2021 the original author or authors.
- *
- * All rights reserved. This program and the accompanying materials are
- * made available under the terms of the Eclipse Public License v2.0 which
- * accompanies this distribution and is available at
- *
- * https://www.eclipse.org/legal/epl-v20.html
- */
-
-package example.configuration;
-
-import org.testng.annotations.AfterClass;
-import org.testng.annotations.Test;
-
-public class FailingAfterClassConfigurationMethodTestCase {
-
-	@AfterClass
-	public void afterClass() {
-		throw new AssertionError("boom");
-	}
-
-	@Test
-	public void test() {
-	}
-}
